common spark handle class
