+++ src/atc/delta/delta_handle.py
@@ -6,7 +6,7 @@ from atc.atc_exceptions import AtcException
 from atc.configurator.configurator import Configurator
 from atc.functions import get_unique_tempview_name, init_dbutils
 from atc.spark import Spark
-from atc.tables.TableHandle import TableHandle
+from atc.tables.SparkHandle import SparkHandle
 from atc.utils.CheckDfMerge import CheckDfMerge
 from atc.utils.GetMergeStatement import GetMergeStatement
 
@@ -23,13 +23,10 @@ class DeltaHandleInvalidFormat(DeltaHandleException):
     pass
 
 
-class DeltaHandle(TableHandle):
+class DeltaHandle(SparkHandle):
     def __init__(self, name: str, location: str = None, data_format: str = "delta"):
-        self._name = name
-        self._location = location
-        self._data_format = data_format
 
-        self._partitioning: Optional[List[str]] = None
+        super().__init__(name, location, data_format)
 
         self._validate()
 
@@ -101,60 +98,6 @@ class DeltaHandle(TableHandle):
         self.drop()
         if self._location:
             init_dbutils().fs.rm(self._location, True)
-
-    def create_hive_table(self) -> None:
-        sql = f"CREATE TABLE IF NOT EXISTS {self._name} "
-        if self._location:
-            sql += f" USING DELTA LOCATION '{self._location}'"
-        Spark.get().sql(sql)
-
-    def recreate_hive_table(self):
-        self.drop()
-        self.create_hive_table()
-
-    def get_partitioning(self):
-        """The result of DESCRIBE TABLE tablename is like this:
-        +-----------------+---------------+-------+
-        |         col_name|      data_type|comment|
-        +-----------------+---------------+-------+
-        |           mycolA|         string|       |
-        |           myColB|            int|       |
-        |                 |               |       |
-        |   # Partitioning|               |       |
-        |           Part 0|         mycolA|       |
-        +-----------------+---------------+-------+
-        but this method return the partitioning in the form ['mycolA'],
-        if there is no partitioning, an empty list is returned.
-        """
-        if self._partitioning is None:
-            # create an iterator object and use it in two steps
-            rows_iter = iter(
-                Spark.get().sql(f"DESCRIBE TABLE {self.get_tablename()}").collect()
-            )
-
-            # roll over the iterator until you see the title line
-            for row in rows_iter:
-                # discard rows until the important section header
-                if row.col_name.strip() == "# Partitioning":
-                    break
-            # at this point, the iterator has moved past the section heading
-            # leaving only the rows with "Part 1" etc.
-
-            # create a list from the rest of the iterator like [(0,colA), (1,colB)]
-            parts = [
-                (int(row.col_name[5:]), row.data_type)
-                for row in rows_iter
-                if row.col_name.startswith("Part ")
-            ]
-            # sort, just in case the parts were out of order.
-            parts.sort()
-
-            # discard the index and put into an ordered list.
-            self._partitioning = [p[1] for p in parts]
-        return self._partitioning
-
-    def get_tablename(self) -> str:
-        return self._name
 
     def upsert(
         self,
